fix: send chat response metadata
