+++ dify/dify-support/dify-support-chat/src/main/java/io/github/guoshiqiufeng/dify/chat/dto/response/ChatMessageSendCompletionResponse.java
@@ -41,89 +41,4 @@ public class ChatMessageSendCompletionResponse extends ChatMessageSendResponse i
 
     @JsonDeserialize(using = CompletionDataDeserializer.class)
     private CompletionData data;
-
-    private Metadata metadata;
-
-
-    @Data
-    public static class Metadata implements Serializable {
-
-        private static final long serialVersionUID = 5739363389399951537L;
-
-        private Usage usage;
-
-        @JsonAlias("retriever_resources")
-        private List<RetrieverResources> retrieverResources;
-
-    }
-
-    @Data
-    public static class RetrieverResources implements Serializable {
-
-        private static final long serialVersionUID = 1023974561298465172L;
-
-        @JsonAlias("position")
-        private Integer position;
-
-        @JsonAlias("dataset_id")
-        private String datasetId;
-
-        @JsonAlias("dataset_name")
-        private String datasetName;
-
-        @JsonAlias("document_id")
-        private String documentId;
-
-        @JsonAlias("document_name")
-        private String documentName;
-
-        @JsonAlias("segment_id")
-        private String segmentId;
-
-        private Float score;
-
-        private String content;
-    }
-
-    @Data
-    public static class Usage implements Serializable {
-
-        private static final long serialVersionUID = -164237129958299138L;
-
-        @JsonAlias("prompt_tokens")
-        private Integer promptTokens;
-
-        @JsonAlias("prompt_unit_price")
-        private String promptUnitPrice;
-
-        @JsonAlias("prompt_price_unit")
-        private String promptPriceUnit;
-
-        @JsonAlias("prompt_price")
-        private String promptPrice;
-
-        @JsonAlias("completion_tokens")
-        private Integer completionTokens;
-
-        @JsonAlias("completion_unit_price")
-        private String completionUnitPrice;
-
-        @JsonAlias("completion_price_unit")
-        private String completionPriceUnit;
-
-        @JsonAlias("completion_price")
-        private String completionPrice;
-
-        @JsonAlias("total_tokens")
-        private Integer totalTokens;
-
-        @JsonAlias("total_price")
-        private String totalPrice;
-
-        @JsonAlias("currency")
-        private String currency;
-
-        @JsonAlias("latency")
-        private Double latency;
-    }
 }
